// lib/api/types.ts
export interface components {
    schemas: {
        Student: {
            student_id: number;
            name: string;
            class: string;
            /** Format: date */
            dob: string;
        };
        HealthRecord: {
            record_id: number;
            student_id: number;
            /** Format: date */
            date: string;
            height: number;
            weight: number;
            notes?: string;
        };
    };
    responses: never;
    parameters: {
        StudentId: number;
        RecordId: number;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}